Work towards V2.0 - Backend stuff
This is the completion of the back end of things towards creating a small OLED screen which shows what song is currently playing on Spotify.
With this backend I have completed the API requests need for the information to be displayed.  Also utilises a config file so that private credentials are not hard coded and allows for others to utilise it.
Now to come up with the code that displays the information to the screen.