Work towards V2.0 - Frontend stuff :tv:
With a massive help from @alhockly the front end of stuff is sorted. With the OLED screen now shows what song is currently playing, the Artist aswell as an animated seek bar showing how far play back is currently.
Thats the bulk of the work done towards #2
Up next is optimising the display and adding a shuffle icon or maybe an icon to show what current playback is on?

Co-Authored-By: Alex Hockly <alhockly@gmail.com>